Create spektral.data module.
This module adds abstractions to represent graphs and graph datasets, as well as data loaders for automatically feeding batches of graphs to a Model (disjoint and batch mode are supported for now).
The commit adds:
- Graph, to represent graphs
- Dataset, to represent graph datasets
- Loader, BatchLoader, DisjointLoader, to feed batches of data in the corresponding formats. If TF 2.4 is installed, the Loaders can be passed directly to Model.fit() by calling Loader.tf().

spektral.utils.data is now merged in spektral.data.utils.
This commit also adds tests for the new module.