Add script for training with large image datasets
Added a script which uses a custom image generator to load images in batches. The script uses a dataset from the Kaggle Competition Plant Seedlings Classification. This is just an example script, for this case flow_from_directory is a better option but a sample only using generator is shown.